+++ src/components/index.js
@@ -5,6 +5,7 @@ export * from './CodeHighlight';
 export * from './DetailsKeyShortcutsWrapper';
 export * from './EditKeyShortcutsWrapper';
 export * from './Callout';
+export * from './FieldAssignedTo';
 export * from './FieldOrganization';
 export * from './ListView';
 export * from './ListTemplate';