+++ src/settings/MappingProfiles/detailsSections/edit/index.js
@@ -3,4 +3,5 @@ export * from './MappingHoldingsDetails';
 export * from './MappingItemDetails';
 export * from './MappingMARCBibDetails';
 export * from './MappingInvoiceDetails';
+export * from './MappingOrderDetails';
 export * from './MappingMARCAuthorityDetails';